Slow down the worker thread when API quota exceded
Currently, individual jobs are postponed in execution, but workers have a full queue that they're constantly trying to send, thus negating the brake system put in place to slow them down when this happens.
This ensures the thread(s) are also waiting, to eliminate the chance that the downloader will ultimately fail. As long as there are not many parallel threads working.